+++ chaos_theory/docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add a `filter` generator modifier to `#[derive(Arbitrary)]` for structs/enums/fields.
 - Rename `Generator::{filter_assume, filter}` to `Generator::{filter, try_filter}`.
 - Rename `make::{from_fn_assume, from_next_assume}` to `make::{from_fn_some, from_next_some}`.
 - Make `make::index` produce `usize` and panic for an empty range; add `make::try_index`.